Set local API ip (emulator only). List of available products is now fetched from the API (Otherwise generate mock data).

+++ ServerConnect/app/src/main/java/com/example/cashmanager/data/model/Cart.kt
@@ -8,9 +8,9 @@ class Cart : Serializable {
     /**
      * Total value of the cart
      */
-    val billTotal: Float
+    val billTotal: Double
         get() {
-            var total = 0f
+            var total = 0.0
             for(product in products)
                 total += product.first.price * product.second
             return total
